+++ Products/DataGridField/skins/DataGridWidget/datecolumn.js
@@ -1,0 +1,4 @@
+$(document).ready(function() {
+  var selector = ".datagridwidget-cell input[type='date']";
+  $(selector).dateinput();
+});